install @nostrability/schemata from github
