+++ packages/common/package.json
@@ -77,8 +77,10 @@
   },
   "devDependencies": {
     "@hirez_io/observer-spy": "^2.2.0",
+    "@nostrability/schemata": "github:nostrability/schemata",
     "@types/debug": "^4.1.12",
     "@types/hash-sum": "^1.0.2",
+    "@types/node": "^22.19.1",
     "ajv": "^8.17.1",
     "ajv-formats": "^3.0.1",
     "nostr-tools": "^2.18",
