--- schemata/nips/nip-01/kind-0/schema.content.yaml
@@ -1,49 +1,0 @@
-$schema: "http://json-schema.org/draft-07/schema#"
-title: "Metadata Event Schema"
-type: "object"
-properties:
-  name:
-    type: "string"
-    minLength: 1
-  about:
-    type: "string"
-  picture:
-    type: "string"
-    format: "uri"
-  display_name:
-    type: "string"
-    description: "An alternative, richer name. 'name' must always be set."
-  website:
-    type: "string"
-    format: "uri"
-    description: "A web URL related to the event author."
-  banner:
-    type: "string"
-    format: "uri"
-    description: "A URL to a wide picture for background display."
-  bot:
-    type: "boolean"
-    description: "Indicates if content is automated."
-  lud16: 
-    type: "string"
-    format: "email"
-  lud06: 
-    type: "string"
-
-  # deprecations
-  displayName:
-    deprecated: true
-    description: "Use 'display_name' instead."
-    errorMessage: {
-      deprecated: "Use 'display_name' instead."
-    }
-  username:
-    deprecated: true
-    description: "Use 'name' instead."
-    errorMessage: {
-      deprecated: "Use 'display_name' instead."
-    }
-  
-required:
-  - "name"
-additionalProperties: true
