--- schemata/nips/nip-01/kind-0/schema.yaml
@@ -1,9 +1,0 @@
-$schema: http://json-schema.org/draft-07/schema#
-title: kind0
-allOf: 
-  - $ref: "@/note.yaml"
-  - type: object
-    properties:
-      kind:
-        const: 0
-        errorMessage: "kind must equal 0"
