--- schemata/nips/nip-01/note/schema.yaml
@@ -1,43 +1,0 @@
-$schema: http://json-schema.org/draft-07/schema
-type: object
-properties:
-  content:
-    type: string
-    errorMessage: "content must be a string"
-    description: "The content of the note"
-  created_at:
-    type: integer
-    errorMessage: "created_at must be a timestamp expressed in seconds (not milliseconds)"
-    description: "The timestamp of the note creation"
-  id:
-    allOf:
-    - $ref: "@/secp256k1.yaml"
-    errorMessage: "id must be a valid hash"
-    description: "The id is a hash derived as specified in NIP-01"
-  kind:
-    type: integer
-  pubkey:
-    allOf:
-    - $ref: "@/secp256k1.yaml"
-    errorMessage: "pubkey must be a secp256k1 public key"
-    description: "The public key of the note's author"
-  sig:
-    type: string
-    pattern: "^[a-f0-9]{128}$"
-    errorMessage: "sig must be 128 lowercase hex characters (64 bytes)"
-    description: "The cryptographic signature of the note"
-  tags:
-    type: array
-    errorMessage: "tags must be an array of valid tag tuples"
-    description: "The tags of the note"
-    items: { $ref: "@/tag.yaml" }
-      
-required:
-- content
-- created_at
-- id
-- kind
-- pubkey
-- sig
-- tags
-additionalProperties: false
