--- schemata/nips/nip-01/secp256k1/schema.yaml
@@ -1,3 +1,0 @@
-$schema: http://json-schema.org/draft-07/schema#    
-type: string
-pattern: "^[a-f0-9]{64}$"
