--- schemata/nips/nip-01/tag/schema.yaml
@@ -1,5 +1,0 @@
-$schema: "http://json-schema.org/draft-07/schema#"
-type: array
-items:
-  type: string
-uniqueItems: false
